Refactor css code v2
