+++ Frontend/adminpanel/src/components/PageCountSwitcher/PageCountSwitcher.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
-import GeneralSelect from './UI/select/GeneralSelect';
-import { PaginationContext } from '../provider/context/PaginationProvider';
+import GeneralSelect from '../UI/select/GeneralSelect';
+import { PaginationContext } from '../../provider/context/PaginationProvider';
 
 export default function PageCountSwitcher() {
