+++ Frontend/adminpanel/src/components/RecordFilter/RecordFilter.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import GeneralInput from './UI/input/GeneralInput';
-import GeneralSelect from './UI/select/GeneralSelect';
+import GeneralInput from '../UI/input/GeneralInput';
+import GeneralSelect from '../UI/select/GeneralSelect';
 
 export default function RecordFilter({ filter, setFilter, fields }) {
