+++ Frontend/adminpanel/src/components/TablesList/TablesList.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import GeneralButton from './UI/button/GeneralButton';
+import GeneralButton from '../UI/button/GeneralButton';
+import cl from './index.module.css'
 
 export default function TablesList({ setCurrentTable, tables, title }) {
 
@@ -13,7 +14,7 @@ export default function TablesList({ setCurrentTable, tables, title }) {
 
     return (
         <div>
-            <h4 className='listHeader'>
+            <h4 className={cl.listHeader}>
                 {title}
             </h4>
             <div className='container'>
