+++ Frontend/adminpanel/src/components/TablesList/index.module.css
@@ -1,0 +1,3 @@
+.listHeader {
+    text-align: center;
+}
